Add files via upload
This is my submission for the Digital Nerds contest. The theme was "What has 2020 taught us?".
For me, the answer is that 2020 showed me how people always have diverging opinions, even contradicting science. So I made a game in which you are a platypus fighting prejudice and fake news!
Please bear through the intro, I ran out of time to make it beautiful.

P.S.: The story is a reference to the "Hitchhiker's Guide to the Galaxy" .